requirements.txt
numpy>=1.21.0
torch>=2.0.0
transformers>=4.30.0
biopython>=1.79
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
requests>=2.26.0
kahip>=3.14